requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
passlib
bcrypt==4.0.1
cryptography
python-dotenv
python-multipart
python-jose[cryptography]
scikit-learn
xgboost
lightgbm
shap
numpy
pandas
scipy
pyarrow
neurokit2
pywavelets
websockets
reportlab
matplotlib